requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-dotenv==1.0.1
google-generativeai==0.3.2
pydantic==2.6.4
python-multipart==0.0.6
docx2txt==0.8
pymupdf==1.24.1
textract==1.6.5
python-docx==1.1.0
# Additional dependencies for textract
antiword==0.37
beautifulsoup4==4.12.3
six==1.16.0